Updated folder structure of the 'slice-codec' crate to be more modular.
