+++ slice-codec/src/decoding/mod.rs
@@ -1,0 +1,7 @@
+// Copyright (c) ZeroC, Inc.
+
+pub mod decode_from;
+pub mod decoder;
+
+// This module is private because it doesn't export any types, just implementations.
+mod decoding;
