+++ slice-codec/src/encoding/mod.rs
@@ -1,0 +1,7 @@
+// Copyright (c) ZeroC, Inc.
+
+pub mod encode_into;
+pub mod encoder;
+
+// This module is private because it doesn't export any types, just implementations.
+mod encoding;
