+++ slicec/src/definition_types.rs
@@ -5,7 +5,8 @@
 
 #![allow(dead_code)]
 
-use slice_codec::buffer::{InputSource, OutputTarget};
+use slice_codec::input_source::InputSource;
+use slice_codec::output_target::OutputTarget;
 use slice_codec::decode_from::DecodeFrom;
 use slice_codec::decoder::Decoder;
 use slice_codec::encode_into::EncodeInto;
